+++ lib_embedding/src/main/java/util/external_software_wrappers/NitpickWrapper.java
@@ -43,9 +43,9 @@ public class NitpickWrapper {
             BufferedReader stdInput = new BufferedReader(new InputStreamReader(proc.getInputStream()));
             BufferedReader stdError = new BufferedReader(new InputStreamReader(proc.getErrorStream()));
             if (!proc.waitFor(timeout + 20, unit)) {
+                ProcessKiller.killProcess(proc);
                 log.fine(filename.toString() + " : Proof Timeout");
                 this.timeout = true;
-                if (proc != null) ProcessKiller.destroyProc(proc, 1500L);
             }else{
                 Instant end = Instant.now();
                 Duration delta = Duration.between(start,end);
@@ -58,25 +58,22 @@ public class NitpickWrapper {
             while ((s = stdError.readLine()) != null) {
                 stderr += s;
             }
-            JavaProcess process = Processes.newJavaProcess(proc);
-            if (process.isAlive()) ProcessKiller.destroyProc(proc, 1500L);
         } catch (IOException e) {
             if (this.stderr == null) this.stderr = e.getMessage();
             if (this.stdout == null) this.stdout = e.getMessage();
-            if (proc != null) ProcessKiller.destroyProc(proc, 1500L);
         } catch (InterruptedException e) {
-            log.fine(filename.toString() + " : Interrupted Exception.");
+            ProcessKiller.killProcess(proc);
+            System.err.println(filename.toString() + " : Interrupted Exception.");
             if (this.stderr == null) this.stderr = e.getMessage();
             if (this.stdout == null) this.stdout = e.getMessage();
             this.timeout = true;
-            if (proc != null) ProcessKiller.destroyProc(proc, 1500L);
         }finally {
             this.status = extractSZSStatus(this.stdout);
             this.durationNitpick = this.extractNitpickDuration();
             //System.out.println(this.status);
-            if (proc != null) ProcessKiller.destroyProc(proc, 1500L);
         }
-        ProcessKiller.killProcess(proc);
+        JavaProcess process = Processes.newJavaProcess(proc);
+        if (process.isAlive()) ProcessKiller.killProcess(proc);
     }
 
     private double extractNitpickDuration(){
